updates delete button to be a modal and deletes practice modal button
